+++ CHANGELOG.md
@@ -4,6 +4,7 @@ master
 * Add --type/t flag to "actions" subcommand
 * Add fiat amount value to "actions" subcommand
 * Can give path to config file via option and environment variable
+* Support automatic detection of SushiSwap trades
 
 0.0.11
 ------